+++ CHANGES.md
@@ -6,6 +6,8 @@
 - Fixes: #214, #217, #218
 - GUI refinements
 - "Renew Session" button and session expiration notification timings revised
+- OpenVPN updated to 2.5.8-20230214
+    - openvpn 2.5.8 1d81df042eae416a4e83e6a433ae2b937c5a10a4
 
 
 ## [3.3.6](https://github.com/Amebis/eduVPN/compare/3.3.5...3.3.6) (2022-11-23)